requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.10
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
djangorestframework==3.14.0
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
